requirements.txt
discord.py==2.2.2
setproctitle==1.3.2
